Relocate Client Server Classes to Client
