+++ vcell-apiclient/src/main/java/org/vcell/api/utils/Auth0ConnectionUtils.java
@@ -1,4 +1,4 @@
-package cbit.vcell.server;
+package org.vcell.api.utils;
 
 import cbit.vcell.resource.PropertyLoader;
 import cbit.vcell.resource.ResourceUtil;
