+++ vcell-client/src/main/java/cbit/vcell/client/DocumentWindowManager.java
@@ -15,9 +15,7 @@ import java.util.Hashtable;
 import javax.swing.JComponent;
 import javax.swing.JPanel;
 
-import org.vcell.util.ObjectNotFoundException;
 import org.vcell.util.document.User;
-import org.vcell.util.document.UserLoginInfo.DigestedPassword;
 import org.vcell.util.document.VCDocument;
 import org.vcell.util.document.VCDocument.VCDocumentType;
 import org.vcell.util.document.VCDocumentInfo;
@@ -28,12 +26,10 @@ import cbit.rmi.event.DataJobEvent;
 import cbit.rmi.event.ExportEvent;
 import cbit.vcell.client.desktop.biomodel.DocumentEditor;
 import cbit.vcell.client.desktop.simulation.SimulationWindow;
-import cbit.vcell.client.server.ConnectionStatus;
+import org.vcell.api.server.ConnectionStatus;
 import cbit.vcell.client.task.AsynchClientTask;
 import cbit.vcell.client.task.ClientTaskDispatcher;
 import cbit.vcell.export.server.ExportSpecs;
-import cbit.vcell.geometry.Geometry;
-import cbit.vcell.geometry.GeometryInfo;
 import cbit.vcell.simdata.OutputContext;
 import cbit.vcell.solver.VCSimulationIdentifier;
 import cbit.xml.merge.XmlTreeDiff;
