+++ vcell-client/src/main/java/cbit/vcell/client/RequestManagerAdapter.java
@@ -8,7 +8,6 @@ import javax.swing.filechooser.FileFilter;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.document.BioModelInfo;
 import org.vcell.util.document.MathModelInfo;
-import org.vcell.util.document.UserLoginInfo.DigestedPassword;
 import org.vcell.util.document.VCDataIdentifier;
 import org.vcell.util.document.VCDocument;
 import org.vcell.util.document.VCDocument.DocumentCreationInfo;
@@ -19,9 +18,9 @@ import cbit.vcell.client.FieldDataWindowManager.DataSymbolCallBack;
 import cbit.vcell.client.TopLevelWindowManager.OpenModelInfoHolder;
 import cbit.vcell.client.data.DataViewerController;
 import cbit.vcell.client.data.MergedDatasetViewerController;
-import cbit.vcell.client.server.AsynchMessageManager;
+import org.vcell.api.server.AsynchMessageManager;
 import cbit.vcell.client.server.ClientServerInfo;
-import cbit.vcell.client.server.ConnectionStatus;
+import org.vcell.api.server.ConnectionStatus;
 import cbit.vcell.client.server.UserPreferences;
 import cbit.vcell.client.task.AsynchClientTask;
 import cbit.vcell.clientdb.DocumentManager;
