+++ vcell-client/src/main/java/cbit/vcell/client/VCellClient.java
@@ -13,8 +13,8 @@ package cbit.vcell.client;
 import cbit.vcell.biomodel.BioModel;
 import cbit.vcell.client.desktop.DocumentWindowAboutBox;
 import cbit.vcell.client.server.ClientServerInfo;
-import cbit.vcell.client.server.ClientServerManager;
-import cbit.vcell.client.server.ClientServerManager.InteractiveContextDefaultProvider;
+import org.vcell.api.server.ClientServerManager;
+import org.vcell.api.server.ClientServerManager.InteractiveContextDefaultProvider;
 import cbit.vcell.client.task.AsynchClientTask;
 import cbit.vcell.client.task.ClientTaskDispatcher;
 import cbit.vcell.desktop.ClientLogin;
@@ -23,7 +23,6 @@ import cbit.vcell.mathmodel.MathModel;
 import cbit.vcell.server.VCellConnectionFactory;
 import com.google.inject.Inject;
 import com.install4j.api.launcher.ApplicationLauncher;
-import org.vcell.util.document.UserLoginInfo.DigestedPassword;
 import org.vcell.util.document.VCDocument;
 import org.vcell.util.document.VCDocument.VCDocumentType;
