+++ vcell-client/src/main/java/cbit/vcell/client/VCellGuiInteractiveContextDefaultProvider.java
@@ -1,6 +1,6 @@
 package cbit.vcell.client;
 
-import cbit.vcell.client.server.ClientServerManager.InteractiveContextDefaultProvider;
+import org.vcell.api.server.ClientServerManager.InteractiveContextDefaultProvider;
 import cbit.vcell.client.server.InteractiveClientServerContext;
 
 public class VCellGuiInteractiveContextDefaultProvider implements InteractiveContextDefaultProvider {
