+++ vcell-core/src/main/java/cbit/vcell/message/server/bootstrap/client/RemoteProxyVCellConnectionFactory.java
@@ -15,7 +15,6 @@ import cbit.vcell.message.VCRpcRequest;
 import cbit.vcell.message.VCellQueue;
 import cbit.vcell.resource.PropertyLoader;
 import cbit.vcell.server.Auth0ConnectionUtils;
-import cbit.vcell.server.ConnectionException;
 import cbit.vcell.server.VCellConnection;
 import cbit.vcell.server.VCellConnectionFactory;
 import com.google.gson.Gson;
@@ -27,8 +26,6 @@ import org.vcell.DependencyConstants;
 import org.vcell.api.client.VCellApiClient;
 import org.vcell.api.client.VCellApiClient.RpcDestination;
 import org.vcell.api.client.VCellApiRpcRequest;
-import org.vcell.api.common.AccessTokenRepresentation;
-import org.vcell.api.common.events.*;
 import org.vcell.restclient.ApiException;
 import org.vcell.restclient.model.AccesTokenRepresentationRecord;
 import org.vcell.util.document.KeyValue;
