+++ vcell-core/src/main/java/cbit/vcell/server/SessionManager.java
@@ -11,12 +11,10 @@
 package cbit.vcell.server;
 
 import cbit.vcell.client.server.AsyncMessageManagerInterface;
-import org.vcell.util.DataAccessException;
-import org.vcell.util.document.User;
-
-import cbit.vcell.client.server.AsynchMessageManager;
 import cbit.vcell.field.io.FieldDataFileOperationResults;
 import cbit.vcell.field.io.FieldDataFileOperationSpec;
+import org.vcell.util.DataAccessException;
+import org.vcell.util.document.User;
 
 /**
  * Insert the type's description here.
